+++ www/astro.config.mjs
@@ -8,11 +8,10 @@ import remarkCodeTitles from "remark-code-titles";
 import rehypeSlug from "rehype-slug";
 import rehypeAutolinkHeadings from "rehype-autolink-headings";
 import sitemap from "@astrojs/sitemap";
-/**
- * @link https://astro.build/config
- */
+
+/** @link https://astro.build/config */
 export default defineConfig({
-  site: `https://create.t3.gg`,
+  site: `https://create.t3.gg/`,
   markdown: {
     remarkPlugins: [remarkCodeTitles],
     rehypePlugins: [
@@ -43,7 +42,6 @@ export default defineConfig({
       theme: "rose-pine",
       wrap: true,
     },
-    extendDefaultPlugins: true,
   },
   integrations: [
     react(),